requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
chromadb
pysqlite3-binary
langchain
langchain-community
langchain-google-genai
langchain-chroma